Move regression test to right unit test file
